Altering coloring and computation of image output to show both internal density of oregen and the y-center of each xz pair inside the vein.
The more "green" the greater the peak density in that area. The more
blue, the closer to the yMin of the vein's "center" is at that x,z --
the more purple, the closer to the yMax of the vein's "center".
Basically, the closer you are to yMin/yMax, the stronger you will
observe the greeness while mining.